Clarify the usage of "image name"
My colleague and I were banging our head against the wall trying get get the `kustomize edit set image` command to work until we realised that we were mixing up the "container name" and the "image name" values. This distinction is hidden when the image name and the container name is the same (e.g. just "nginx" as we saw in other blog posts). It might help to highlight that in the documentation, as proposed.