Minor grammar nits
There is widespread confusion around the use of apostrophes for pluralization. The rules governing this are really a hot mess, which leads to many misuses. But if you pay careful attention to well-edited documents, you'll almost never see it used. The best rule of thumb is, you never really _need_ them, so it's best to reword your sentences to avoid them entirely. There are some valid cases for pluralization with apostrophes, but the whole thing is a general mess so is better avoided when possible. See https://www.dailywritingtips.com/when-to-form-a-plural-with-an-apostrophe/ for a good treatment of the topic. In this document's example, it's best to just spell out the words for clarity.